long read isoform reconstruction and quantification

This function takes bam file of genomic alignments and performs isoform recontruction and gene and transcript expression quantification. It also allows saving of read class files of alignments, extending provided annotations, and quantification based on extended annotations. When multiple samples are provided, extended annotations will be combined across samples to allow comparison.

bambu(
  reads = NULL,
  readClass.file = NULL,
  readClass.outputDir = NULL,
  annotations = NULL,
  genomeSequence = NULL,
  stranded = FALSE,
  ncore = 1,
  yieldSize = NULL,
  isoreParameters = NULL,
  emParameters = NULL,
  extendAnnotations = TRUE,
  verbose = FALSE
)

Arguments

reads
A string or a vector of strings specifying the paths of bam files for genomic alignments, or a BamFile object or a BamFileList object (see Rsamtools).

readClass.file
A string or a vector of strings specifying the read class files that are saved during previous run of bambu.

readClass.outputDir
A string variable specifying the path to where read class files will be saved.

annotations
A TxDb object or A GRangesList object obtained by prepareAnnotations or prepareAnnotationsFromGTF.

genomeSequence
A fasta file or a BSGenome object.

ncore
specifying number of cores used when parallel processing is used, defaults to 1.

yieldSize
see Rsamtools.

isoreParameters
A list of controlling parameters for isoform reconstruction process:

prefix specifying prefix for new gene Ids (genePrefix.number), defaults to empty

remove.subsetTx indicating whether filter to remove read classes which are a subset of known transcripts(), defaults to TRUE

min.readCount specifying minimun read count to consider a read class valid in a sample, defaults to 2

min.readFractionByGene specifying minimum relative read count per gene, highly expressed genes will have many high read count low relative abundance transcripts that can be filtered, defaults to 0.05

min.sampleNumber specifying minimum sample number with minimum read count, defaults to 1

min.exonDistance specifying minum distance to known transcript to be considered valid as new, defaults to 35

min.exonOverlap specifying minimum number of bases shared with annotation to be assigned to the same gene id, defaults 10 base pairs

emParameters
A list of controlling parameters for quantification algorithm estimation process:

maxiter specifying maximum number of run interations, defaults to 10000.

bias specifying whether to correct for bias, defaults to FALSE.

conv specifying the covergence trheshold control, defaults to 0.0001.

extendAnnotations
A logical variable indicating whether annotations are to be extended for quantification.

verbose
A logical variable indicating whether processing messages will be printed.

Value

A list of two SummarizedExperiment object for transcript expression and gene expression.

Details

Main function

Examples