plot.bambu

plotSEOuptut

plotBambu(
    se,
    group.variable = NULL,
    type = c("annotation", "pca", "heatmap"),
    gene_id = NULL,
    transcript_id = NULL
)

Arguments

se
An summarized experiment object obtained from bambu or transcriptToGeneExpression.

group.variable
Variable for grouping in plot, has be to provided if choosing to plot PCA.

type
plot type variable, a values of annotation for a single gene with heatmap for isoform expressions, pca, or heatmap, see details.

gene_id
specifying the gene_id for plotting gene annotation, either gene_id or transcript_id has to be provided when type = "annotation".

transcript_id
specifying the transcript_id for plotting transcript annotation, either gene_id or transcript_id has to be provided when type = "annotation"

Value

A heatmap plot for all samples

Details

type indicates the type of plots to be plotted. There are two types of plots can be chosen, PCA or heatmap.

Examples

se <- readRDS(system.file("extdata",
"seOutputCombined_SGNex_A549_directRNA_replicate5_run1_chr9_1_1000000.rds",
package = "bambu"))
plotBambu(se, type = "PCA")